+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # VXG Mobile SDK iOS
 
+### MobileSDK 2.0.82.220204
+
+ CloudSDK
+Version 2.0.82_20220204:
+- added new Refresh Token API
+- added new advanced settings: 
+  reconnect on http error for hls sources.
+  default value: 404
+
+ PlayerSDK
+Version 2.0.20220112:
+- fixed the problem with passing 64-bit pointers to ffmpeg by dictionary
+- fixed AAC ASC gegenrated config from stream info
+- added configuration setting for generate ASC. Default: 0
+- added new advanced settings: reconnect on http error. Default: 404
+
+ EncoderSDK
+Version 2.0.20220110:
+
 ### MobileSDK 2.0.80.211118
 
  CloudSDK